// src/types/api.ts
export interface paths {
    "/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["AppController_getHello"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/evento": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["eventoController_findAll"];
        put: operations["eventoController_update"];
        post: operations["eventoController_create"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/evento/{nombre}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["eventoController_findOne"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/evento/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post?: never;
        delete: operations["eventoController_remove"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/tipo-entrada": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["TipoEntradaController_findAll"];
        put?: never;
        post: operations["TipoEntradaController_create"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/tipo-entrada/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["TipoEntradaController_findOne"];
        put?: never;
        post?: never;
        delete: operations["TipoEntradaController_remove"];
        options?: never;
        head?: never;
        patch: operations["TipoEntradaController_update"];
        trace?: never;
    };
    "/ticket": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["TicketController_findAll"];
        put?: never;
        post: operations["TicketController_create"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/ticket/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["TicketController_findOne"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/cliente": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["ClienteController_findAll"];
        put?: never;
        post: operations["ClienteController_create"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/cliente/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["ClienteController_findOne"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/staff": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["StaffController_findAll"];
        put?: never;
        post: operations["StaffController_create"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/staff/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["StaffController_findOne"];
        put?: never;
        post?: never;
        delete: operations["StaffController_remove"];
        options?: never;
        head?: never;
        patch: operations["StaffController_update"];
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        CreateTipoEntradaDto: {
            nombre: string;
            precio: number;
            cantidadEntradas: number;
        };
        CreateEventoDto: {
            nombre: string;
            /** Format: date-time */
            fecha: string;
            descripcion: string;
            tipoEntrada: components["schemas"]["CreateTipoEntradaDto"];
            creadorId: string;
            imagePublicId: string;
        };
        returnEventoDto: {
            id: string;
            nombre: string;
            /** Format: date-time */
            fecha: string;
            descripcion: string;
            tipoEntrada: components["schemas"]["CreateTipoEntradaDto"][];
            creadorId: string;
            imagePublicId: string;
        };
        ReturnTipoEntrada: {
            nombre: string;
            precio: number;
            cantidadEntradas: number;
        };
        UpdateTipoEntradaDto: {
            nombre?: string;
            precio?: number;
            cantidadEntradas?: number;
        };
        updateTipoEntradaDto: {
            _id: string;
            nombre?: string;
            descripcion?: string;
            /** Format: date-time */
            fecha?: string;
            tipoEntrada?: components["schemas"]["UpdateTipoEntradaDto"];
        };
        CreateTicketDto: Record<string, never>;
        CreateClienteDto: Record<string, never>;
        CreateStaffDto: Record<string, never>;
        UpdateStaffDto: Record<string, never>;
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    AppController_getHello: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
    eventoController_findAll: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["returnEventoDto"][];
                };
            };
        };
    };
// ...
}